Rename root package to `sbot`
