+++ controllers/token_scorer/token_scorer.py
@@ -13,7 +13,7 @@ REPO_ROOT = Path(__file__).resolve().parent.parent.parent
 
 sys.path.insert(1, str(REPO_ROOT / 'modules'))
 
-from sr.robot.utils import get_robot_device  # isort:skip
+from sbot.utils import get_robot_device  # isort:skip
 from shared_utils import Owner, TOKENS, TargetInfo, TargetType  # isort:skip
 import controller_utils  # isort:skip
