+++ modules/sbot/arduino_devices.py
@@ -6,9 +6,9 @@ from controller import (
     TouchSensor,
     DistanceSensor as WebotsDistanceSensor,
 )
-from sr.robot.utils import map_to_range, get_robot_device
-from sr.robot.randomizer import add_jitter
-from sr.robot.output_frequency_limiter import OutputFrequencyLimiter
+from sbot.utils import map_to_range, get_robot_device
+from sbot.randomizer import add_jitter
+from sbot.output_frequency_limiter import OutputFrequencyLimiter
 
 LOGGER = logging.getLogger(__name__)
