+++ modules/sbot/vision/angles.png
