+++ modules/sbot/vision/image.py
